requirements.txt
mysql-connector-python==8.0.26
protobuf==3.17.3
six==1.16.0
